Make pathfinding slightly more robust to inconsistent data
I found a case where name synonyms working inconsistently meant that,
in the ADS data, B was a coauthor with A, but A was not a coauthor
with B. This meant the path-finding iteration went on for one loop
than it should, and route of distance N and N+1 both existed between
src and dest. This caused dist_from_src in some nodes to jump by
2 between neighbors, rather than 1.

This change ensures the most direct route is found, regardless of
slight inconsistencies such as that in the final graph.