Prevent the type of inconsistency mentioned in the previous commit
Ensures the graph is walked in what's more of a breadth-first way.
Using a stack rather than a queue was interacting with the "only visit
a node once" logic to cause inaccurate distance values to be propagated
through from one point in the graph to the end of the graph.